refactor: move _apply_post_dispersion_pixel_response to kl_pipe/grism.py
New kl_pipe/grism.py hosts the post-dispersion 2D-image stage of the
grism pipeline (parallel to kl_pipe/dispersion.py for the cube->2D step).
Mechanical move: function body unchanged; import path updated at three
callsites (SourceModel.render_grism, KLModel.render_grism, test_cube_psf).

Frees kl_pipe/model.py of grism-specific plumbing ahead of KLModel
deletion in commit 10, where the helper would otherwise be left
orphaned. No behavior change; 66 grism-touching tests pass.

Phase 3 commit 6.
